feat: Customizable UI with addon xk6 extension
Embedded UI can be customized with separated xk6 extension
